Added support for full connection string passed to constructor or connection string name
Allows a full connection string to be passed to the constructor while
still supporting the connection string name from config file. This
allows for environments or situations where you don't use the a config
file. Also renamed ConnectionString to _connectionString to be inlined
with visible naming conventions in the file, which is also in line with
common conventions for C#.