Null exception if value is null
Changed double if statements in ValidatesPresenceOf into an if/else
statement, because if the value is null it would through a
NullReferenceException when it called ToString on it.

In ValidateIsCurrency added a return after value == null check is true
as we don't need to process the rest of the method and if you did and
value was null he ToString call would throw a NullReferenceException.