Insert Bug Fix
Insert Tests contain a test of the original code vs the modified version
with a dummy table based on the Employee table of the Northwind
database. Did not include the app.config with the localconnection
connection string.

The original Insert command has been renamed to InsertOriginal and the
fixed version that has been added to the Massive class file is named
InsertNew.