Set fetches are now 12-15% faster
The RecordToExpando method was kind of slow as it obtained the value
twice and also not using the bulk fetch method of the datareader. It is
still not a fast performer due to Expando and the boxing/unboxing which
eventually happens but at least this bottleneck is out of the way.